+++ victorious/VictoriousIOSSDK/Models/FanLoyalty.swift
@@ -9,14 +9,14 @@
 import SwiftyJSON
 
 public struct FanLoyalty {
-    public let points: Int64
+    public let points: Int
     public let level: Int
     public let progress: Int
 }
 
 extension FanLoyalty {
     public init?(json: JSON) {
-        guard let points = json["points"].int64,
+        guard let points = json["points"].int,
             let level = json["level"].int,
             let progress = json["progress"].int else {
                 return nil
